add phase existence by id validator
