+++ libs/Profio.Application/Phases/Commands/CreatePhaseCommand.cs
@@ -10,7 +10,7 @@ namespace Profio.Application.Phases.Commands;
 
 public sealed record CreatePhaseCommand : CreateCommandBase
 {
-  public int Order { get; set; } = 0;
+  public int Order { get; set; } = 1;
   public string? RouteId { get; set; }
 };
 sealed class CreatePhaseCommandHandler : CreateCommandHandlerBase<CreatePhaseCommand, Phase>
@@ -24,7 +24,7 @@ public sealed class CreatePhaseCommandValidator : AbstractValidator<CreatePhaseC
   public CreatePhaseCommandValidator(RouteExistenceByIdValidator routeValidator)
   {
     RuleFor(x => x.Order)
-      .GreaterThanOrEqualTo(0);
+      .GreaterThanOrEqualTo(1);
 
     RuleFor(x => x.RouteId)
       .SetValidator(routeValidator!);
